+++ src/contents/settings/exclusions.tsx
@@ -1,6 +1,5 @@
 import DirExclusionsItem from "@/components/settings-item/dir-exclusions";
-import PuaExclusionsItem from "@/components/settings-item/pua-exclusions";
-import { useBackendSettings } from "@/hooks/use-settings";
+import { useExclusions } from "@/hooks/use-settings";
 import { DEFAULT_BACKEND_SETTINGS } from "@/lib/settings";
 import { BackendSettings } from "@/lib/types/settings";
 import ExclusionsLoader from "@/loaders/components/exclusions";
@@ -9,41 +8,29 @@ import { toast } from "sonner";
 
 export default function ExclusionsSettings(){
      const [isFetching, startTransition] = useTransition()
-     const {fetchBackendSettings, setBackendSettings} = useBackendSettings()
+     const {getExclusions, setExclusions: setExclusionData} = useExclusions()
      const [exclusions, setExclusions] = useState<BackendSettings["exclusions"]>(DEFAULT_BACKEND_SETTINGS.exclusions);
      useEffect(()=>{
           startTransition(async()=>{
                try {
-                    const settings = await fetchBackendSettings("exclusions")
-                    setExclusions(val=>!settings ? val : settings)
+                    const stored = await getExclusions()
+                    setExclusions(val=>!stored ? val : stored)
                } catch (err){
-                    toast.error("Failed to fetch existing protection settings");
+                    toast.error("Failed to fetch existing exclusions");
                     console.error(err)
                }
           })
      },[])
-     const updateExclusions = async <K extends keyof BackendSettings["exclusions"]>(key: K, value: BackendSettings["exclusions"][K]) => {
-          await setBackendSettings("exclusions",key,value);
-          setExclusions(prev=>({...prev, [key]: value}))
+     const updateExclusions = async(value: BackendSettings["exclusions"]) => {
+          await setExclusionData(value);
+          setExclusions(value)
      }
-     const dirExclusions = useMemo(()=>!exclusions.directory ? DEFAULT_BACKEND_SETTINGS.exclusions.directory : exclusions.directory,[exclusions.directory]);
-     const puaExclusions = useMemo(()=>!exclusions.puaCategory ? DEFAULT_BACKEND_SETTINGS.exclusions.puaCategory : exclusions.puaCategory,[exclusions.puaCategory]);
-     const handleExclusionAction = async  <K extends keyof BackendSettings["exclusions"]>(value: string, key: K, action: "exclude" | "remove") => {
-          let newArr: string[] = []
-          if(key === "directory") {
-               const {directory} = exclusions
-               const mainArr = !directory ? DEFAULT_BACKEND_SETTINGS.exclusions.directory : directory
-               newArr = action==="exclude" ? [...mainArr,value] : mainArr.filter(val=>val!==value)
-          } else {
-               const {puaCategory} = exclusions
-               const mainArr = !puaCategory ? DEFAULT_BACKEND_SETTINGS.exclusions.puaCategory : puaCategory
-               newArr = action==="exclude" ? [...mainArr,value] : mainArr.filter(val=>val!==value)
-          }
-          await updateExclusions(key,newArr);
-          setExclusions(prev=>({
-               ...prev,
-               [key]: newArr
-          }))
+     const dirExclusions = useMemo(()=>!exclusions ? DEFAULT_BACKEND_SETTINGS.exclusions: exclusions,[exclusions]);
+     const handleExclusionAction = async (value: string, action: "exclude" | "remove") => {
+          if(value.trim()==="") return;
+          const mainArr = !exclusions ? DEFAULT_BACKEND_SETTINGS.exclusions : exclusions
+          const newArr = action==="exclude" ? [...mainArr,value] : mainArr.filter(val=>val!==value)
+          await updateExclusions(newArr);
      }
      return (
           <div className="px-1 py-2 space-y-3">
@@ -52,17 +39,8 @@ export default function ExclusionsSettings(){
                ) : (
                     <DirExclusionsItem
                          data={dirExclusions}
-                         onSubmit={values=>handleExclusionAction(values.path,"directory","exclude")}
-                         onDelete={path=>handleExclusionAction(path,"directory","remove")}
-                    />
-               )}
-               {isFetching ? (
-                    <ExclusionsLoader items={puaExclusions.length}/>
-               ) : (
-                    <PuaExclusionsItem
-                         data={puaExclusions}
-                         onSubmit={values=>handleExclusionAction(values.category,"puaCategory","exclude")}
-                         onDelete={category=>handleExclusionAction(category,"puaCategory","remove")}
+                         onSubmit={values=>handleExclusionAction(values.path,"exclude")}
+                         onDelete={path=>handleExclusionAction(path,"remove")}
                     />
                )}
           </div>