+++ src/loaders/settings/exclusions.tsx
@@ -4,7 +4,6 @@ export default function ExclusionsSettingsLoader(){
      return (
           <div className="px-1 py-2 space-y-3 w-full">
                <ExclusionsLoader/>
-               <ExclusionsLoader/>
           </div>
      )
 }